bulk backport from trunk (rev 6389 + 6390 + 6391 + 6392 + 6397 + 6398 + 6399): - fixed bogus inline declaration - a function prototype cannot be inline -> reported by suncc - varios fixes reported by succ : 	- useless returns 	- return val in void functions - fixed include path for unixodbc lib for OpenSolaris. - fixed bug in testing the success of the parse_to() function - fixed all makefiles using autodetection based on "which" -> older version was not working on non-linux systems. - fixed makefiles for the perl modules ; the syntax of makefile (for calling external shell commands) is not like in shell script  :) . - for gcc added some extra compile flags to get rid of the "redundant declaration" warning due auto created C files - fixed solaris and suncc compiling
  Credits go to Michal Bielicki @ Seventh Signal




git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/branches/1.6@6400 689a6050-402a-0410-94f2-e92a70836424

+++ modules/tm/t_hooks.c
@@ -42,7 +42,7 @@
 struct tmcb_head_list* req_in_tmcb_hl = 0;
 
 struct tmcb_head_list tmcb_pending_hl = {0,0};
-unsigned int tmcb_pending_id = -1;
+unsigned int tmcb_pending_id = (unsigned int)-1;
 
 
 void empty_tmcb_list(struct tmcb_head_list *head)
